Initial Commit
This is a project that I worked on in the past attempting to Recreate the Controls and Camera of StarCraft 2. I aborted the project with minimal work however I am restarting it for my Practicum 2 for Game 330. I am revising everything I did before and did some before I put it in git. Part of it uses an RTS tutorial which I no longer have access to, but it the code should be far enough removed by now before my changes that it will be fine.